+++ Sources/KeyboardKit/Status/KeyboardStatus.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardState.swift
+//  KeyboardStatus.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2024-03-26.
@@ -8,8 +8,9 @@
 
 import Foundation
 
-/// This namespace contains state-related types.
+/// This namespace contains keyboard status-related types to
+/// handle the enabled state of a keyboard.
 ///
 /// The namespace doesn't contain protocols, open classes or
 /// types of higher importance.
-public struct KeyboardState {}
+public struct KeyboardStatus {}